refactor report2.py and server.py
